API测试

+++ app/src/main/java/com/xuejian/client/lxp/module/toolbox/im/ContactlistFragment.java
@@ -100,7 +100,7 @@ public class ContactlistFragment extends Fragment {
                     // 进入群聊列表页面
                     startActivity(new Intent(getActivity(), GroupsActivity.class));
                 } else {
-                    startActivity(new Intent(getActivity(), HisMainPageActivity.class).putExtra("userId", adapter.getItem(position).getUserId().intValue()));
+                    startActivity(new Intent(getActivity(), HisMainPageActivity.class).putExtra("userId", adapter.getItem(position).getUserId()));
                     getActivity().overridePendingTransition(R.anim.slide_in_from_right, R.anim.slide_out_to_left);
                 }
             }
